+++ src/Parser.h
@@ -61,7 +61,7 @@ private:
 #if AZARAC_ENABLE_NANKAI
     // Process Nankai Trough page aggregation
     // Returns true if message should be output (page aggregation complete or not Nankai)
-    bool processNankaiAggregation(const Message& decoded, Message& out, uint64_t current_ms);
+    bool processNankaiAggregation(const Message& decoded, Message& out, const Mt43Data* d, uint64_t current_ms);
 #endif
 };